Added SSH class, changed to ssh-copy-id
