+++ lib/clearsight/version.rb
@@ -1,3 +1,3 @@
 module Clearsight
-  VERSION = "1.1.0"
+  VERSION = "1.2.0.rc1"
 end
